Day 8 parts 1 and 2
This was the first day I needed a hint.
I figured out part 1 no problem, but part 2 was running indefinitely.
Had to go searching to figure out that the inputs all have paths that
are complete cycles, where the starting node is on that cycle, so
finding the numnber of steps for each path and calculating the least
common multiple will give you the total number of steps.