Add neighbor counts during initial roll population
This way, when we remove a roll, we can decrement all of its neighboring
rolls, and then to see if the neighbor can be removed, it's a semple comparison,
rather than checking all of its neighbors.
And while we're add it, we build the initial removal queue in the
read_rolls function too,
since both parts will need that right away.